requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.0
mysql-connector-python==8.0.33
aioredis==2.0.1
